+++ packages/edit-site/src/components/dataviews-actions/index.js
@@ -4,11 +4,14 @@ import { useMemo } from '@wordpress/element';
 import { privateApis as routerPrivateApis } from '@wordpress/router';
 import { useDispatch, useSelect } from '@wordpress/data';
 import { store as coreStore } from '@wordpress/core-data';
+import { privateApis as editorPrivateApis } from '@wordpress/editor';
+import { viewPostRevisions } from '@wordpress/fields';
 import { addQueryArgs } from '@wordpress/url';
 import { PATTERN_TYPES } from '../../utils/constants';
 import { unlock } from '../../lock-unlock';
 
 const { useLocation, useHistory } = unlock( routerPrivateApis );
+const { usePostActions } = unlock( editorPrivateApis );
 
 export const useSetActiveTemplateAction = () => {
 	const activeTheme = useSelect( ( select ) =>
@@ -89,6 +92,44 @@ export const useEditPostAction = () => {
 			},
 		} ),
 		[ history ]
+	);
+};
+
+/**
+ * Returns the actions registered for a post type, with the site editor specific
+ * overrides applied.
+ *
+ * @param {Object} options The options passed to `usePostActions`.
+ * @return {Array} The actions.
+ */
+export const useSiteEditorPostActions = ( options ) => {
+	const postTypeActions = usePostActions( options );
+	const history = useHistory();
+	return useMemo(
+		() =>
+			postTypeActions.map( ( action ) =>
+				action.id === viewPostRevisions.id
+					? {
+							...action,
+							callback( items ) {
+								const post = items[ 0 ];
+								history.navigate(
+									addQueryArgs(
+										`/${ post.type }/${ post.id }`,
+										{
+											canvas: 'edit',
+											revision:
+												post._links?.[
+													'predecessor-version'
+												]?.[ 0 ]?.id,
+										}
+									)
+								);
+							},
+					  }
+					: action
+			),
+		[ postTypeActions, history ]
 	);
 };