+++ packages/edit-site/CHANGELOG.md
@@ -29,6 +29,7 @@
 ### Bug Fixes
 
 -   Identity: Decode HTML entities in the Site Title and Site Tagline fields. ([#81269](https://github.com/WordPress/gutenberg/pull/81269))
+-   Open the "View revisions" action in the site editor revisions view instead of the `revision.php` screen. ([#81366](https://github.com/WordPress/gutenberg/pull/81366))
 
 ## 7.1.0 (2026-07-29)